Replace dataset fetcher with external library
